+++ packages/client/src/components/base/Pagination.tsx
@@ -15,6 +15,7 @@ interface PaginateProps {
   className?: string;
   hasMore?: boolean;
   loading?: boolean;
+  justifyBetween?: boolean;
 }
 
 const Paginate: React.FC<PaginateProps> = ({
@@ -25,48 +26,65 @@ const Paginate: React.FC<PaginateProps> = ({
   hasMore = false,
   loading = false,
   className,
+  justifyBetween = false,
 }) => {
   if (urlF) {
     return (
-      <Flexbox direction="row" gap="2" alignItems="center" className={className}>
+      <Flexbox
+        direction="row"
+        gap="2"
+        alignItems="center"
+        justify={justifyBetween ? 'between' : 'start'}
+        className={className}
+      >
         <Text semibold className="mr-4">{`Page ${active + 1} of ${count}`}</Text>
-        <Button color="primary" href={urlF(0)} disabled={active === 0 || loading}>
-          <MoveToStartIcon size={16} />
-        </Button>
-        <Button color="primary" href={urlF(active - 1)} disabled={active === 0 || loading}>
-          <ArrowLeftIcon size={16} />
-        </Button>
-        <Button color="primary" href={urlF(active + 1)} disabled={active === count - 1 || loading}>
-          <ArrowRightIcon size={16} />
-        </Button>
-        <Button color="primary" href={urlF(count - 1)} disabled={active === count - 1 || loading}>
-          <MoveToEndIcon size={16} />
-        </Button>
+        <Flexbox direction="row" gap="2" alignItems="center">
+          <Button color="primary" href={urlF(0)} disabled={active === 0 || loading}>
+            <MoveToStartIcon size={16} />
+          </Button>
+          <Button color="primary" href={urlF(active - 1)} disabled={active === 0 || loading}>
+            <ArrowLeftIcon size={16} />
+          </Button>
+          <Button color="primary" href={urlF(active + 1)} disabled={active === count - 1 || loading}>
+            <ArrowRightIcon size={16} />
+          </Button>
+          <Button color="primary" href={urlF(count - 1)} disabled={active === count - 1 || loading}>
+            <MoveToEndIcon size={16} />
+          </Button>
+        </Flexbox>
       </Flexbox>
     );
   }
 
   if (onClick) {
     return (
-      <Flexbox direction="row" gap="2" alignItems="center" className={className}>
+      <Flexbox
+        direction="row"
+        gap="2"
+        alignItems="center"
+        justify={justifyBetween ? 'between' : 'start'}
+        className={className}
+      >
         <Text semibold className="mr-4">{`Page ${active + 1} of ${count}${hasMore ? '+' : ''}`}</Text>
-        {loading && <Spinner sm />}
-        <Button color="primary" onClick={() => onClick(0)} disabled={active === 0 || loading}>
-          <MoveToStartIcon size={16} />
-        </Button>
-        <Button color="primary" onClick={() => onClick(active - 1)} disabled={active === 0 || loading}>
-          <ArrowLeftIcon size={16} />
-        </Button>
-        <Button
-          color="primary"
-          onClick={() => onClick(active + 1)}
-          disabled={(active === count - 1 && hasMore === false) || loading}
-        >
-          <ArrowRightIcon size={16} />
-        </Button>
-        <Button color="primary" onClick={() => onClick(count - 1)} disabled={active === count - 1 || loading}>
-          <MoveToEndIcon size={16} />
-        </Button>
+        <Flexbox direction="row" gap="2" alignItems="center">
+          {loading && <Spinner sm />}
+          <Button color="primary" onClick={() => onClick(0)} disabled={active === 0 || loading}>
+            <MoveToStartIcon size={16} />
+          </Button>
+          <Button color="primary" onClick={() => onClick(active - 1)} disabled={active === 0 || loading}>
+            <ArrowLeftIcon size={16} />
+          </Button>
+          <Button
+            color="primary"
+            onClick={() => onClick(active + 1)}
+            disabled={(active === count - 1 && hasMore === false) || loading}
+          >
+            <ArrowRightIcon size={16} />
+          </Button>
+          <Button color="primary" onClick={() => onClick(count - 1)} disabled={active === count - 1 || loading}>
+            <MoveToEndIcon size={16} />
+          </Button>
+        </Flexbox>
       </Flexbox>
     );
   }