+++ CHANGES.md
@@ -6,11 +6,17 @@ Since 1.5.0
 - In Cube table view, Added the card item hover effect to each item in the group when hovering over the card group header
 - Added a shareable link to generate a random sample pack each time
 - allow urls in cube compare input
+- Refactored dashboard mobile UX
+- Added "View More" links to explore page
 
 # Bug Fixes
 
 - Fixed a bug where deck dropdowns weren't working
 - Fixed a bug where cube exports didn't use sort/filter correctly
 - Fix transforming card full name (both faces) not able to bulk import
+- Fixed a bug where the cube search index sort would fall out of sync with reality
+- Fixed a bug where the card tags on hover were being displayed on mobile
 
 # Technical Changes
+
+- Service now runs in a VPC